API: fix story_details_by_location
This API endpoint gives us a list of stories, and for each, it finds the list of sites associated with that story and and gives the distance to the closest site and the name of the suburb the site is in. However, previously the code to get the "nearest" suburb just did this:

MIN(site.suburb)

which gives you the suburb that comes first alphabeticallly.

This commit adds a bunch of SQL to fix that. It's really quite absurd how much code was added for that one little thing; if you can find a better way of doing it, pull requests welcome...

(To be fair, the functions added might be useful for other things as well perhaps.)