Adds a StoryScreen to the Location so we can view either intro or body.
When users first are shown a story, they should see the intro. Later they
might click through to the body. This change just changes the data model
to enable that, but doesn't actually change any of view code. So
the code below introduces the new `type`, and the rest of the diff is
just updating signatures to make elm compile.

types.elm:

 type Location id =
       Discovering
-    | Viewing id ItemView
+    | Viewing id ItemView StoryScreen
     | ViewingFavourites

+type StoryScreen = Intro | Body