about-screen content area should have background so map stays hidden
for some reason, Safari is calculating .about-screen div height as height of screen and only applying background as far as that. If we put background on .content-area inside .about-screen, the problem goes away. (relates to issue #143, "Text bg seems buggy and doesn't extend to the bottom of the text body on mobile& narrow browsers.")